// frontend/src/services/api/models/SeatLayoutPassenger.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SeatLayoutPassenger = {
    seat_id?: number;
    seat_number?: string;
    seat_type?: string;
    price?: number;
    status?: 'AVAILABLE' | 'BOOKED';
};